refactor(identifications): move bulk creation logic to models_future
The bulk endpoint's resolve/validate/save loop now lives in
ami/main/models_future/identifications.py, keeping the viewset action thin:
parse the request, authorize the batch, delegate, shape the response. This
follows the models_future pattern (filters, occurrence, projects) and gives
the follow-up model-level agreement validation (#1378) a natural home.

Also, per review feedback, the per-item save handler catches IntegrityError
instead of the broader DatabaseError so connection-level failures surface as
errors instead of being reported as routine per-item rejections, and the
per-item error message no longer claims the occurrence was deleted when the
cause may be another related record. Inline comments are trimmed to the
repo's short-comment norm.
